+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.0.3
 
 * Fixed `Color.toARGB32()` not working on Flutter 3.27.3 by adding an extension.
+* Updated inline documentation.
 
 ## 0.0.2